Upgrade to ESLint v9
