+++ package.json
@@ -4,7 +4,7 @@
   "description": "Module for MagicMirror². Displays the Merriam-Webster word of the day along with its part of speech, pronunciation, definition, and an example.",
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/jmwyds/MMM-WordOfTheDay"
+    "url": "https://github.com/jmwyds/MMM-WordOfTheDay"
   },
   "keywords": [
     "MagicMirror",
@@ -17,12 +17,12 @@
   "author": "jmwyds",
   "license": "MIT",
   "devDependencies": {
-    "@eslint/js": "latest",
-    "@stylistic/eslint-plugin": "latest",
-    "eslint": "^8.56.0",
-    "eslint-plugin-import": "latest",
-    "eslint-plugin-unicorn": "latest",
-    "prettier": "^3.1.1"
+    "@eslint/js": "^9.17.0",
+    "@stylistic/eslint-plugin": "^2.12.1",
+    "eslint": "^9.17.0",
+    "eslint-plugin-import": "^2.31.0",
+    "eslint-plugin-unicorn": "^56.0.1",
+    "prettier": "^3.4.2"
   },
   "scripts": {
     "lint": "eslint . && prettier --check .",
